[TASK] Implement root folder selector and rendition context
Summary:
This change refactors the driver so it supports an CMIS
OperationContext and creates such a context if one is not
provided, allowing the default thumbnail rendition to be
used as alternative representation.

Also splits the folder selector into a repository and
folder selector to allow using a secondary CMIS repository
as file storage. And finally, adjusts the driver to use
and check these values to prevent exceptions from an
unconfigured FAL driver configuration record.

Reviewers: dkd-egerer, #[t]_forgetit-dev

Reviewed By: dkd-egerer, #[t]_forgetit-dev

Differential Revision: https://phabricator.dkd.de/D1198