[TASK] Add default empty items to the driver configuration selector fields
Summary:
When the driver is configured the first item of the dropdown was selected
automatically. This could be a little bit misleading as it looks like this is already
used/saved but is not. So a default item with an empty value has been added which is
selected by default.

Reviewers: #[t]_forgetit-dev, claus.due

Reviewed By: #[t]_forgetit-dev, claus.due

Differential Revision: https://phabricator.dkd.de/D1340