[TASK] Move session initialization from "initialize" to "processConfiguration"
Summary:
Move the code from the `CMISFilesystemDriver::initialize` function to
the `CMISFilesystemDriver::processConfiguration` function to prevent exceptions
are thrown to the users.
All Runtime Exceptions are now catched and the Storeage is temporarly marked as
offline. Also a "nice" error is shown to the user if something went wrong while
initializing the connection.

Test Plan: Open the Storage configuraiton if the configured CMIS Service is not available

Reviewers: #[t]_forgetit-dev, claus.due

Reviewed By: #[t]_forgetit-dev, claus.due

Subscribers: claus.due

Differential Revision: https://phabricator.dkd.de/D1342