[BUGFIX] Check if option value is empty instead of if option key exists
Summary:
The function `CMISFilesystemDriver::getOption()` checks if
an configuration option is set and returns the default value instead.
This check should check if the option is `empty` instead of if the
option exists in the configuration array as this would be true in
most cases and will return an unexpected empty value.

Reviewers: claus.due, #[t]_forgetit-dev

Reviewed By: claus.due, #[t]_forgetit-dev

Differential Revision: https://phabricator.dkd.de/D1414