[BUGFIX] Fix storage id access in SubResolvingDriver::getFolderInfoByIdentifier
Summary:
In `SubResolvingDriver::getFolderInfoByIdentifier` the id of the storage
is returned as part of the result array. The current implementation accessed an
invalid `$this->storageUid` which results in an empty value.
The `CMISFilesystemDriver` class has been extended to provide a method to get the
storage id which is not used.

Reviewers: claus.due, #[t]_forgetit-dev

Reviewed By: claus.due, #[t]_forgetit-dev

Differential Revision: https://phabricator.dkd.de/D1409