[TASK] Emulate CMIS getByPath error
Summary:
This emulates an error that would be thrown by the
CMIS client when attempting to find an object by
path and that object or any part of the path to it
does not exist. The error can then be caught by
other code and result in automatic folder/file creation
when this is required (for example user_upload folder).

Reviewers: dkd-goslar

Reviewed By: dkd-goslar

Differential Revision: https://phabricator.dkd.de/D1803